gNOI System Reboot changes
